Split save game.
